requirements.txt
Flask==2.3.3
requests==2.31.0
gunicorn
groq
newsapi-python
python-dotenv
pytest==7.4.2
pytest-cov==4.1.0
pytest-flask==1.2.0